theme fixes for OS X (installing, search icons)
I had to update the theme install part of setup.py to include directories,
since search engines and their icons are in directories in themes now.  Themes
were still loaded on OS X, so I didn't have to make any changes for that code.
I added code to controls.py to look for custom search icons in the theme rather
than in the Resources directory.

Refs #11056


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@8919 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 720a698c54453ede76e5690a54ee98fc1c618b96

+++ tv/platform/osx/setup.py
@@ -185,11 +185,14 @@ class Config (object):
                     extract = False
                     break            
             if extract:
-                data = themeArchive.read(entry)
                 path = os.path.join(target, entry)
-                f = open(path, "wt")
-                f.write(data)
-                f.close()
+                if entry.endswith('/'):
+                    os.makedirs(path)
+                else:
+                    data = themeArchive.read(entry)
+                    f = open(path, "wt")
+                    f.write(data)
+                    f.close()
         themeArchive.close()
     
     def get_icon_file(self):
@@ -530,10 +533,16 @@ class MiroBuild (py2app):
             print "    (all skipped, already bundled)"
         else:
             os.makedirs(targetDir)
-            for f in glob(os.path.join(sourceDir, "*")):
-                dest = os.path.join(targetDir, os.path.basename(f))
-                print "    %s" % dest
-                shutil.copy(f, dest)
+            def copyDirectory(d):
+                for f in glob(os.path.join(d, "*")):
+                    dest = os.path.join(targetDir, f[len(sourceDir)+1:])
+                    print "    %s" % dest
+                    if os.path.isfile(f):
+                        shutil.copy(f, dest)
+                    elif os.path.isdir(f):
+                        os.makedirs(dest)
+                        copyDirectory(f)
+            copyDirectory(sourceDir)
     
     def clean_up_incomplete_lproj(self):
         print "Wiping out incomplete lproj folders"
